# services/core/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateTaskRequest:
    properties:
      description:
        type: string
      title:
        type: string
    required:
    - title
    type: object
  dto.CreateTaskResponse:
    properties:
      id:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  dto.ListSkillsResponse:
    properties:
      skills:
        items:
          $ref: '#/definitions/dto.SkillResponse'
        type: array
    type: object
...
  dto.SkillResponse:
    properties:
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  dto.TaskResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      plan_id:
        type: string
      position:
        type: integer
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.UpdateTaskRequest:
    properties:
      description:
        type: string
      title:
        type: string
    type: object
  dto.UpdateTaskStatusRequest:
    properties:
      status:
        type: string
    required:
    - status
    type: object
...
host: localhost:8080
info:
  contact:
    name: Mandzhonson
  description: REST API для SkillTracker.
  title: SkillTracker API
  version: "1.0"
paths:
...
  /employee/skills:
    get:
      consumes:
      - application/json
      description: Возвращает список навыков текущего сотрудника
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListSkillsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить список навыков сотрудника
      tags:
      - Employee
  /employee/tasks/{task_id}/status:
    patch:
      consumes:
      - application/json
      description: Изменяет статус задачи (для сотрудника)
      parameters:
      - description: ID задачи
        in: path
        name: task_id
        required: true
        type: string
      - description: Новый статус
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTaskStatusRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить статус задачи
      tags:
      - Employee
...
  /manager/plans/{plan_id}/tasks:
    post:
      consumes:
      - application/json
      description: Создает новую задачу в плане
      parameters:
      - description: ID плана
        in: path
        name: plan_id
        required: true
        type: string
      - description: Данные для создания задачи
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateTaskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: План архивирован
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать задачу
      tags:
      - Tasks
...
  /manager/tasks/{task_id}:
    delete:
      consumes:
      - application/json
      description: Удаляет задачу
      parameters:
      - description: ID задачи
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить задачу
      tags:
      - Tasks
    get:
      consumes:
      - application/json
      description: Возвращает детальную информацию о задаче
      parameters:
      - description: ID задачи
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить задачу по ID
      tags:
      - Tasks
    patch:
      consumes:
      - application/json
      description: Обновляет название и описание задачи
      parameters:
      - description: ID задачи
        in: path
        name: task_id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить задачу
      tags:
      - Tasks
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"